requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
pydantic
redis
